+++ CHANGELOG.rst
@@ -46,6 +46,8 @@ Full list of changes:
 * ``PageObjectRegistry`` is replaced with ``RulesRegistry``; its API is changed:
 
     * **backwards incompatible** dict-like API is removed;
+    * **backwards incompatible** *O(1)* lookups using
+      ``.search(use=PagObject)`` has become *O(N)*;
     * ``search_overrides`` method is renamed to ``search``;
     * ``get_overrides`` method is renamed to ``get_rules``;
     * ``from_override_rules`` method is deprecated;